Create Bokeh_MOVEUAS_V1
comments still need to be updated. there is a bug where if you check at least one box then uncheck all of them there is still a drop down box left on the screen. same for the tables. This runs properly 100% of the time locally and is a little iffy when trying internet access. It also only allows one device to connect to the page at a time, or one device per port. I will fix that at a later date.